Modernize build system for devkitARM r65 / GCC 14
Build fixes:
- libunds: replace with system libnds (stub build target)
- libelm: fix newlib syscall signatures (int fd -> void* fd)
- libelm: st_spare1 -> st_spare4[0]
- akloader/akmenu4: remove -mno-fpu (dropped in GCC)
- akloader/akmenu4: fix specs paths to use system ds_arm7/9.specs
- akloader/akmenu4: -lunds7/-lunds9 -> -lnds7/-lnds9
- FIFO_PACK_VALUE32: add compat define (removed from libnds)
- CARD_DATA_RD -> CARD_DATA_READY
- CARD_COMMAND: add compat define in rpgprotocol.h
- arm7/9destination/executeAddress: cast void* <-> u32
- diropen/dirnext/dirclose: replace with POSIX opendir/readdir/closedir
- timer.h: const -> constexpr for double member
- crtsmall.cpp: mark exception stubs __attribute__((weak))
- crtsmall.cpp: std::nothrow_t default init (C++17)
- patches_pr.cpp: s8 -> u8 for lookup table with values > 127
- savemngr.cpp: sizeof cast to u32 for std::min
- popmenu: u32 -> size_t return type match
- stricmp -> strcasecmp
- loader/dsrom/romloader: arm7destination void* casts

+++ woodrpg/akloader/arm9/source/loader.cpp
@@ -120,7 +120,7 @@ bool loadFile(const char* filename,u32& originSpeed,u32& changedSpeed,u32& sizeM
   }
 
   // load_arm7
-  if( __NDSHeader->arm7destination >= 0x037F8000 || 0x23232323 == gamecode(__NDSHeader->gameCode) ) //23->'#'
+  if( (u32)__NDSHeader->arm7destination >= 0x037F8000 || (u32)0x23232323 == gamecode(__NDSHeader->gameCode) ) //23->'#'
   {
     u32 arm7_code=__NDSHeader->arm7romOffset;
     u32 arm7_size=__NDSHeader->arm7binarySize;
